+++ examples/menu.rs
@@ -23,7 +23,7 @@ impl SimpleComponent for AppModel {
     view! {
         #[root]
         main_window = gtk::ApplicationWindow {
-            set_title: Some("Simple app"),
+            set_title: Some("Menu example"),
             set_default_width: 300,
             set_default_height: 100,
             gtk::Box {