+++ relm4/benches/stress_test.rs
@@ -33,7 +33,7 @@ impl SimpleComponent for AppModel {
 
     fn init_root() -> Self::Root {
         gtk::Window::builder()
-            .title("Simple app")
+            .title("Stress test benchmark")
             .default_width(300)
             .default_height(100)
             .build()